requirements.txt
absl-py==1.4.0
accelerate==1.12.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.18.4
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
apsw==3.51.2.0
apswutils==0.1.2
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
array_record==0.8.3
arrow==1.4.0
asttokens==3.0.1
astunparse==1.6.3
attrs==25.4.0
audioread==3.1.0
backcall==0.2.0
beartype==0.22.9
beautifulsoup4==4.13.5
betterproto==2.0.0b6
bleach==6.3.0
blinker==1.4
blis==1.3.3
catalogue==2.0.10
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
chex==0.1.90
click==8.3.1
cloudpathlib==0.23.0
cloudpickle==3.1.2
colab-ssh==0.3.27
colorlog==6.10.1
confection==0.1.5
contourpy==1.3.3
cryptography==3.4.8
cycler==0.12.1
cymem==2.0.13
datasets==4.5.0
dbus-python==1.2.18
debugpy==1.8.15
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.0
distrax==0.1.5
distro==1.9.0
dm-tree==0.1.9
docopt==0.6.2
docstring_parser==0.17.0
einops==0.8.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
entrypoints==0.4
etils==1.13.0
executing==2.2.1
fastai==2.8.7
fastcore==1.12.14
fastdownload==0.0.7
fastjsonschema==2.21.2
fastlite==0.2.4
fastprogress==1.1.5
fasttransform==0.0.2
filelock==3.24.2
flatbuffers==25.12.19
flax==0.11.2
fonttools==4.61.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.10.0
gast==0.7.0
gdown==5.2.1
google==3.0.0
google-api-core==2.29.0
google-auth==2.47.0
google-auth-oauthlib==1.2.4
google-cloud-core==2.5.0
google-cloud-storage==3.9.0
google-crc32c==1.8.0
google-genai==1.63.0
google-pasta==0.2.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grain==0.2.15
greenlet==3.3.2
grpcio==1.78.0
grpclib==0.4.9
gspread==6.2.1
gspread-dataframe==4.0.0
h11==0.16.0
h2==4.3.0
h5py==3.15.1
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.20.2
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.4.1
humanize==4.15.0
hyperframe==6.1.0
idna==3.11
ImageIO==2.37.2
immutabledict==4.3.1
importlib-metadata==4.6.4
importlib_resources==6.5.2
iniconfig==2.3.0
ipykernel==6.17.1
ipyparallel==8.8.0
ipython==8.12.3
ipython-genutils==0.2.0
ipywidgets==7.7.1
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.7.2
jaxlib==0.7.2
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.1.6
joblib==1.5.3
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter_client==7.4.9
jupyter_core==5.9.1
jupyter_kernel_gateway @ git+https://github.com/googlecolab/kernel_gateway@b134e9945df25c2dcb98ade9129399be10788671
jupyter_server==2.14.0
jupyter_server_terminals==0.5.4
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.16
kaggle==1.7.4.5
kagglehub==0.3.13
keras==3.10.0
keyring==23.5.0
kiwisolver==1.4.9
lark==1.3.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
llvmlite==0.46.0
Mako==1.1.3
Markdown==3.3.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.0
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.2.0
ml_dtypes==0.5.4
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
multiprocess==0.70.18
murmurhash==1.0.15
namex==0.1.0
narwhals==2.16.0
nbclassic==1.3.3
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.1
notebook==6.5.7
notebook_shim==0.2.4
numba==0.63.1
numpy==2.4.2
oauthlib==3.3.1
opt_einsum==3.4.0
optax==0.2.7
optree==0.18.0
optuna==4.7.0
orbax-checkpoint==0.11.32
overrides==7.7.0
packaging==26.0
pandas==2.2.2
pandas-stubs==2.2.2.240909
pandocfilters==1.5.1
parso==0.8.6
pexpect==4.9.0
pickleshare==0.7.5
pillow==12.1.1
platformdirs==4.9.2
plotly==5.24.1
pluggy==1.6.0
plum-dispatch==2.6.1
pooch==1.9.0
portpicker==1.5.2
preshed==3.0.12
prometheus_client==0.24.1
promise==2.3
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
psutil==5.9.5
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.3.0
pyparsing==3.3.2
PySocks==1.7.1
pytest==8.4.2
python-apt==0.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-fasthtml==0.12.42
python-json-logger==4.0.0
python-multipart==0.0.22
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.3
pyzmq==26.2.1
referencing==0.37.0
regex==2026.1.15
requests==2.32.4
requests-oauthlib==2.0.0
requirements-parser==0.9.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.3.2
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-image==0.25.2
scikit-learn==1.8.0
scipy==1.16.3
seaborn==0.13.2
SecretStorage==3.3.1
Send2Trash==2.1.0
sentencepiece==0.2.1
setuptools==75.2.0
shap==0.50.0
shellingham==1.5.4
simple-parsing==0.1.8
simplejson==3.20.2
six==1.17.0
sklearn-pandas==2.2.0
slicer==0.0.8
smart_open==7.5.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.8.3
soxr==1.0.0
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.46
srsly==2.5.2
stack-data==0.6.3
starlette==0.52.1
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
tensorflow-datasets==4.9.9
tensorflow-metadata==1.17.3
tensorflow-probability==0.25.0
tensorstore==0.1.81
termcolor==3.3.0
terminado==0.18.1
text-unidecode==1.3
thinc==8.3.10
threadpoolctl==3.6.0
tifffile==2026.2.16
tinycss2==1.4.0
tokenizers==0.22.2
toml==0.10.2
toolz==1.1.0
torch==2.9.0
tornado==6.5.1
tpu-info==0.8.1
tqdm==4.67.3
traitlets==5.7.1
transformers==5.0.0
treescope==0.1.10
typer==0.24.0
typer-slim==0.24.0
types-pytz==2025.2.0.20251108
types-setuptools==80.9.0.20251223
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.41.0
uvloop==0.22.1
vega-datasets==0.9.0
wadllib==1.3.6
wasabi==1.1.3
watchfiles==1.1.1
wcwidth==0.6.0
weasel==0.4.3
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.6
wheel==0.46.3
widgetsnbextension==3.6.10
wrapt==2.1.1
xgboost==3.2.0
xxhash==3.6.0
yarg==0.1.9
yarl==1.22.0
zipp==3.23.0